separated Core in order to add ClI
