+++ src/Core/api/ReturnTypeCreator.h
