+++ src/Core/logic/IDGeneratorInterface.h
@@ -5,7 +5,7 @@
 #ifndef TODOLIST_IDGENERATORINTERFACE_H
 #define TODOLIST_IDGENERATORINTERFACE_H
 
-#include "api/TaskID.h"
+#include "Core/api/TaskID.h"
 
 /*
  *  Class-generator of ID
