+++ src/Core/logic/ViewByDate.cpp
