+++ src/Core/logic/ViewByDate.h
