+++ src/Core/logic/ViewByPriority.cpp
