+++ src/main.cpp
@@ -1,9 +1,5 @@
 
-#include "api/TaskService.h"
-#include "api/TaskServiceCreator.h"
-
 int main() {
-   auto ts = todo_list::createService();
 
    return 0;
 }
