+++ tests/Core/api/TaskServiceTest.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
-#include "api/TaskService.h"
+#include "Core/api/TaskService.h"
 
 using ::testing::Return;
