+++ tests/Core/model/TaskStorageTest.cpp
@@ -3,7 +3,7 @@
 //
 
 #include <gtest/gtest.h>
-#include "model/TaskStorage.h"
+#include "Core/model/TaskStorage.h"
 
 const Task t1 = Task::create(
         boost::gregorian::date(2020, 7, 31),
